Add active_bugzilla/testing/fake_bugzilla_server.rb
Adds a testing helper class, that can be included as needed to handle
simulating a running bugzilla's xmlrpc web service.

It requires manually configuring responses, for example:

  bug_search_response = {
    "bugs" => [
      { "id" => 1234567, "description" => "doesn't work!!" }
      { "id" => 7654321, "description" => "super slow!!1!"  }
    ]
  }

  FakeBugzillaServer.start_with :response => bug_search_response do
    action "Bug.search" do
      response
    end
  end

  # code running against server...

  FakeBugzillaServer.stop

But allows the exercising of the XMLRPC::Client code, so when swapping
out different parsers, there is better transparency on whether or not
things work properly still.