[WIP][DONOTMERGE] Alternative FakeBugzillaServer handlers
This is an alternative approach that I first took when specing this out
that probably shouldn't be used, but is a different way of handling the
test data.

I moved away from this approach since it ended up being pretty static
with it's test data, and would have required a lot of data duplication
for the approach we had for our specs.

The current implementation does require a bit more code within the
specs (which will be made apparent in future commits), but I think it
ends up being less work overall.  This approach, while probably more
code, would allow moving the example responses outside of the spec
files, which may or may not be a good thing.

* * *

I most likely will rebase out this commit, but putting it in here for
now as an example of something I tried and way to prompt further
discussion.